requirements.txt
# LangChain core and community packages
langchain==0.3.27
langchain-core==0.3.76
langchain-community==0.3.29
langchain-text-splitters==0.3.11
langchain-groq==0.3.8
langchain-huggingface==0.3.1
langsmith==0.4.27 

# Embeddings and ML
sentence-transformers==5.1.0
transformers==4.56.1
faiss-cpu==1.12.0
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.3.2

# PDF handling
pypdf==3.14.0

# Web / API
fastapi==0.111.1
uvicorn[standard]==0.23.2
requests==2.32.5
python-dotenv==1.1.1

# Frontend
streamlit==1.49.1
